+++ Projects/MQTT_DeviceControl/mqtthandler.h
@@ -10,7 +10,7 @@ class MqttHandler : public QObject
   Q_PROPERTY(QMqttClient::ClientState state READ state WRITE setState NOTIFY stateChanged FINAL)
   Q_PROPERTY(QString temperature READ temperature WRITE setTemperature NOTIFY temperatureChanged FINAL)
   Q_PROPERTY(QString humidity READ humidity WRITE setHumidity NOTIFY humidityChanged FINAL)
-  Q_PROPERTY(quint16 slider READ slider WRITE setSlider NOTIFY sliderChanged FINAL)
+  Q_PROPERTY(quint32 slider READ slider WRITE setSlider NOTIFY sliderChanged FINAL)
   Q_PROPERTY(bool led READ led WRITE setLed NOTIFY ledChanged FINAL)
 public:
   explicit MqttHandler(QObject *parent = nullptr);
@@ -30,8 +30,8 @@ public:
   QString humidity() const;
   void setHumidity(const QString &newHumidity);
 
-  quint16 slider() const;
-  void setSlider(quint16 newSlider);
+  quint32 slider() const;
+  void setSlider(quint32 newSlider);
 
   bool led() const;
   void setLed(bool newLed);
@@ -59,7 +59,7 @@ private:
   QMqttClient m_client;
   QString m_temperature;
   QString m_humidity;
-  quint16 m_slider;
+  quint32 m_slider;
   bool m_led;
 };
